requirements.txt
# HeinerCast Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# File handling
aiofiles>=23.2.0
python-magic>=0.4.27

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Encryption
cryptography>=41.0.0

# Google Drive (optional)
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.115.0

# Templates
jinja2>=3.1.3

# Utilities
python-dotenv>=1.0.0
orjson>=3.9.0

# FFmpeg wrapper
ffmpeg-python>=0.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.1.0
isort>=5.13.0
